Spirte class orient; with Ani ext; need adjust drawing priority
