--- test.ts
@@ -1,815 +1,0 @@
-let map = tiles.createTilemap(hex`1000100001010101010101010101010101010101010000000000000000000000000000010100000000000000000000000000000101000000000000000000000000000001010000000000000000000000000000010100000000000000000000000000000101000000000000000000000000000001010000000000000000000000000000010100000000000000000000000000000101000000000000000000000000000001010000000000000000000000000000010100000000000000000000000000000101000000000000000000000000000001010000000000000000000000000000010100000000000000000000000000000101010101010101010101010101010101`, img`
-2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 
-2 . . . . . . . . . . . . . . 2 
-2 . . . . . . . . . . . . . . 2 
-2 . . . . . . . . . . . . . . 2 
-2 . . . . . . . . . . . . . . 2 
-2 . . . . . . . . . . . . . . 2 
-2 . . . . . . . . . . . . . . 2 
-2 . . . . . . . . . . . . . . 2 
-2 . . . . . . . . . . . . . . 2 
-2 . . . . . . . . . . . . . . 2 
-2 . . . . . . . . . . . . . . 2 
-2 . . . . . . . . . . . . . . 2 
-2 . . . . . . . . . . . . . . 2 
-2 . . . . . . . . . . . . . . 2 
-2 . . . . . . . . . . . . . . 2 
-2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 
-`, [image.create(16,16), sprites.castle.tileGrass2], TileScale.Sixteen);
-
-namespace SpriteKind {
-    export const Friend = SpriteKind.create()
-}
-/**
- * sprites operations
- */
-sprites.onOverlap(SpriteKind.Friend, SpriteKind.Projectile, function (sprite, otherSprite) {
-    music.knock.play()
-    otherSprite.destroy()
-    sprite.destroy()
-})
-sprites.onOverlap(SpriteKind.Friend, SpriteKind.Food, function (sprite, otherSprite) {
-    music.pewPew.play()
-    otherSprite.destroy()
-})
-sprites.onOverlap(SpriteKind.Friend, SpriteKind.Friend, function (sprite, otherSprite) {
-    // music.smallCrash.play()
-    if (sprite != sprHero && otherSprite != sprHero) {
-        return
-    }
-    if (otherSprite == sprHero) {
-        otherSprite = sprite
-    }
-    // otherSprite.follow(sprHero, 5, 110)
-})
-// ////////////////////// end engine operations /////////////////////////
-sprites.onOverlap(SpriteKind.Player, SpriteKind.Food, function (sprite, otherSprite) {
-    music.baDing.play()
-    info.changeScoreBy(1)
-    otherSprite.destroy()
-})
-scene.onHitWall(SpriteKind.Projectile, function (sprite, location) {
-    sprite.destroy()
-})
-let ms = 0
-let sprs: Sprite[] = []
-let TileMapMode = true
-game.stats = true
-scene.setBackgroundImage(img`
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    ................................................................................................................................................................
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
-    `)
-// ////////////////////// engine operations /////////////////////////
-tiles.setCurrentTilemap(map)
-// animations, see animations.ts for description in detail
-
-let sprHero = new XYZAniSprite(10, 8, 15, 0, SpriteKind.Friend, sprites.castle.heroFrontAttack1)
-sprs.push(sprHero)
-scene.cameraFollowSprite(sprHero)
-let sprPriness = new XYZAniSprite(10, 7, 15, 0, SpriteKind.Friend, sprites.castle.princessFront0)
-sprs.push(sprPriness)
-const st = new State(6, 6, defaultFov)
-for (let spr of sprs) {
-    spr.setBounceOnWall(true)
-}
-
-character.loopFrames(
-    sprHero,
-    [img`
-    . . . . . . f f f f . . . . . . 
-    . . . . f f f 2 2 f f f . . . . 
-    . . . f f f 2 2 2 2 f f f . . . 
-    . . f f f e e e e e e f f f . . 
-    . . f f e 2 2 2 2 2 2 e e f . . 
-    . . f e 2 f f f f f f 2 e f . . 
-    . . f f f f e e e e f f f f . . 
-    . f f e f b f 4 4 f b f e f f . 
-    . f e e 4 1 f d d f 1 4 e e f . 
-    . . f e e d d d d d d e e f . . 
-    . . . f e e 4 4 4 4 e e f . . . 
-    . . e 4 f 2 2 2 2 2 2 f 4 e . . 
-    . . 4 d f 2 2 2 2 2 2 f d 4 . . 
-    . . 4 4 f 4 4 5 5 4 4 f 4 4 . . 
-    . . . . . f f f f f f . . . . . 
-    . . . . . f f . . f f . . . . . 
-    `, img`
-        . . . . . . . . . . . . . . . .
-        . . . . . . f f f f . . . . . .
-        . . . . f f f 2 2 f f f . . . .
-        . . . f f f 2 2 2 2 f f f . . .
-        . . f f f e e e e e e f f f . .
-        . . f f e 2 2 2 2 2 2 e e f . .
-        . f f e 2 f f f f f f 2 e f f .
-        . f f f f f e e e e f f f f f .
-        . . f e f b f 4 4 f b f e f . .
-        . . f e 4 1 f d d f 1 4 e f . .
-        . . . f e 4 d d d d 4 e f e . .
-        . . f e f 2 2 2 2 e d d 4 e . .
-        . . e 4 f 2 2 2 2 e d d e . . .
-        . . . . f 4 4 5 5 f e e . . . .
-        . . . . f f f f f f f . . . . .
-        . . . . f f f . . . . . . . . .
-    `, img`
-    . . . . . . f f f f . . . . . . 
-    . . . . f f f 2 2 f f f . . . . 
-    . . . f f f 2 2 2 2 f f f . . . 
-    . . f f f e e e e e e f f f . . 
-    . . f f e 2 2 2 2 2 2 e e f . . 
-    . . f e 2 f f f f f f 2 e f . . 
-    . . f f f f e e e e f f f f . . 
-    . f f e f b f 4 4 f b f e f f . 
-    . f e e 4 1 f d d f 1 4 e e f . 
-    . . f e e d d d d d d e e f . . 
-    . . . f e e 4 4 4 4 e e f . . . 
-    . . e 4 f 2 2 2 2 2 2 f 4 e . . 
-    . . 4 d f 2 2 2 2 2 2 f d 4 . . 
-    . . 4 4 f 4 4 5 5 4 4 f 4 4 . . 
-    . . . . . f f f f f f . . . . . 
-    . . . . . f f . . f f . . . . . 
-    `, img`
-    . . . . . . . . . . . . . . . . 
-    . . . . . . f f f f . . . . . . 
-    . . . . f f f 2 2 f f f . . . . 
-    . . . f f f 2 2 2 2 f f f . . . 
-    . . f f f e e e e e e f f f . . 
-    . . f e e 2 2 2 2 2 2 e f f . . 
-    . f f e 2 f f f f f f 2 e f f . 
-    . f f f f f e e e e f f f f f . 
-    . . f e f b f 4 4 f b f e f . . 
-    . . f e 4 1 f d d f 1 4 e f . . 
-    . . e f e 4 d d d d 4 e f . . . 
-    . . e 4 d d e 2 2 2 2 f e f . . 
-    . . . e d d e 2 2 2 2 f 4 e . . 
-    . . . . e e f 5 5 4 4 f . . . . 
-    . . . . . f f f f f f f . . . . 
-    . . . . . . . . . f f f . . . . 
-    `],
-    150,
-    character.rule(Predicate.MovingDown)
-)
-character.loopFrames(
-    sprHero,
-    [img`
-    . . . . . . f f f f . . . . . . 
-    . . . . f f e e e e f f . . . . 
-    . . . f e e e f f e e e f . . . 
-    . . f f f f f 2 2 f f f f f . . 
-    . . f f e 2 e 2 2 e 2 e f f . . 
-    . . f e 2 f 2 f f 2 f 2 e f . . 
-    . . f f f 2 2 e e 2 2 f f f . . 
-    . f f e f 2 f e e f 2 f e f f . 
-    . f e e f f e e e e f e e e f . 
-    . . f e e e e e e e e e e f . . 
-    . . . f e e e e e e e e f . . . 
-    . . e 4 f f f f f f f f 4 e . . 
-    . . 4 d f 2 2 2 2 2 2 f d 4 . . 
-    . . 4 4 f 4 4 4 4 4 4 f 4 4 . . 
-    . . . . . f f f f f f . . . . . 
-    . . . . . f f . . f f . . . . . 
-    `, img`
-    . . . . . . . . . . . . . . . . 
-    . . . . . . f f f f . . . . . . 
-    . . . . f f e e e e f f . . . . 
-    . . . f e e e f f e e e f . . . 
-    . . . f f f f 2 2 f f f f . . . 
-    . . f f e 2 e 2 2 e 2 e f f . . 
-    . . f e 2 f 2 f f f 2 f e f . . 
-    . . f f f 2 f e e 2 2 f f f . . 
-    . . f e 2 f f e e 2 f e e f . . 
-    . f f e f f e e e f e e e f f . 
-    . f f e e e e e e e e e e f f . 
-    . . . f e e e e e e e e f . . . 
-    . . . e f f f f f f f f 4 e . . 
-    . . . 4 f 2 2 2 2 2 e d d 4 . . 
-    . . . e f f f f f f e e 4 . . . 
-    . . . . f f f . . . . . . . . . 
-    `, img`
-    . . . . . . f f f f . . . . . . 
-    . . . . f f e e e e f f . . . . 
-    . . . f e e e f f e e e f . . . 
-    . . f f f f f 2 2 f f f f f . . 
-    . . f f e 2 e 2 2 e 2 e f f . . 
-    . . f e 2 f 2 f f 2 f 2 e f . . 
-    . . f f f 2 2 e e 2 2 f f f . . 
-    . f f e f 2 f e e f 2 f e f f . 
-    . f e e f f e e e e f e e e f . 
-    . . f e e e e e e e e e e f . . 
-    . . . f e e e e e e e e f . . . 
-    . . e 4 f f f f f f f f 4 e . . 
-    . . 4 d f 2 2 2 2 2 2 f d 4 . . 
-    . . 4 4 f 4 4 4 4 4 4 f 4 4 . . 
-    . . . . . f f f f f f . . . . . 
-    . . . . . f f . . f f . . . . . 
-    `, img`
-    . . . . . . . . . . . . . . . . 
-    . . . . . . f f f f . . . . . . 
-    . . . . f f e e e e f f . . . . 
-    . . . f e e e f f e e e f . . . 
-    . . . f f f f 2 2 f f f f . . . 
-    . . f f e 2 e 2 2 e 2 e f f . . 
-    . . f e f 2 f f f 2 f 2 e f . . 
-    . . f f f 2 2 e e f 2 f f f . . 
-    . . f e e f 2 e e f f 2 e f . . 
-    . f f e e e f e e e f f e f f . 
-    . f f e e e e e e e e e e f f . 
-    . . . f e e e e e e e e f . . . 
-    . . e 4 f f f f f f f f e . . . 
-    . . 4 d d e 2 2 2 2 2 f 4 . . . 
-    . . . 4 e e f f f f f f e . . . 
-    . . . . . . . . . f f f . . . . 
-    `],
-    150,
-    character.rule(Predicate.MovingUp)
-)
-character.loopFrames(
-    sprHero,
-    [img`
-    . . . . . . . . . . . . . . . . 
-    . . . . . . f f f f f f . . . . 
-    . . . . f f e e e e f 2 f . . . 
-    . . . f f e e e e f 2 2 2 f . . 
-    . . . f e e e f f e e e e f . . 
-    . . . f f f f e e 2 2 2 2 e f . 
-    . . . f e 2 2 2 f f f f e 2 f . 
-    . . f f f f f f f e e e f f f . 
-    . . f f e 4 4 e b f 4 4 e e f . 
-    . . f e e 4 d 4 1 f d d e f . . 
-    . . . f e e e 4 d d d d f . . . 
-    . . . . 4 d d e 4 4 4 e f . . . 
-    . . . . e d d e 2 2 2 2 f . . . 
-    . . . . f e e f 4 4 5 5 f f . . 
-    . . . . f f f f f f f f f f . . 
-    . . . . . f f . . . f f f . . . 
-    `, img`
-    . . . . . . f f f f f f . . . . 
-    . . . . f f e e e e f 2 f . . . 
-    . . . f f e e e e f 2 2 2 f . . 
-    . . . f e e e f f e e e e f . . 
-    . . . f f f f e e 2 2 2 2 e f . 
-    . . . f e 2 2 2 f f f f e 2 f . 
-    . . f f f f f f f e e e f f f . 
-    . . f f e 4 4 e b f 4 4 e e f . 
-    . . f e e 4 d 4 1 f d d e f . . 
-    . . . f e e e 4 d d d d f . . . 
-    . . . . f f e e 4 4 4 e f . . . 
-    . . . . . 4 d d e 2 2 2 f . . . 
-    . . . . . e d d e 2 2 2 f . . . 
-    . . . . . f e e f 4 5 5 f . . . 
-    . . . . . . f f f f f f . . . . 
-    . . . . . . . f f f . . . . . . 
-    `, img`
-    . . . . . . . . . . . . . . . . 
-    . . . . . . f f f f f f . . . . 
-    . . . . f f e e e e f 2 f . . . 
-    . . . f f e e e e f 2 2 2 f . . 
-    . . . f e e e f f e e e e f . . 
-    . . . f f f f e e 2 2 2 2 e f . 
-    . . . f e 2 2 2 f f f f e 2 f . 
-    . . f f f f f f f e e e f f f . 
-    . . f f e 4 4 e b f 4 4 e e f . 
-    . . f e e 4 d 4 1 f d d e f . . 
-    . . . f e e e e e d d d f . . . 
-    . . . . . f 4 d d e 4 e f . . . 
-    . . . . . f e d d e 2 2 f . . . 
-    . . . . f f f e e f 5 5 f f . . 
-    . . . . f f f f f f f f f f . . 
-    . . . . . f f . . . f f f . . . 
-    `, img`
-    . . . . . . f f f f f f . . . . 
-    . . . . f f e e e e f 2 f . . . 
-    . . . f f e e e e f 2 2 2 f . . 
-    . . . f e e e f f e e e e f . . 
-    . . . f f f f e e 2 2 2 2 e f . 
-    . . . f e 2 2 2 f f f f e 2 f . 
-    . . f f f f f f f e e e f f f . 
-    . . f f e 4 4 e b f 4 4 e e f . 
-    . . f e e 4 d 4 1 f d d e f . . 
-    . . . f e e e 4 d d d d f . . . 
-    . . . . f f e e 4 4 4 e f . . . 
-    . . . . . 4 d d e 2 2 2 f . . . 
-    . . . . . e d d e 2 2 2 f . . . 
-    . . . . . f e e f 4 5 5 f . . . 
-    . . . . . . f f f f f f . . . . 
-    . . . . . . . f f f . . . . . . 
-    `],
-    150,
-    character.rule(Predicate.MovingRight)
-)
-character.loopFrames(
-    sprHero,
-    [img`
-    . . . . f f f f f f . . . . . . 
-    . . . f 2 f e e e e f f . . . . 
-    . . f 2 2 2 f e e e e f f . . . 
-    . . f e e e e f f e e e f . . . 
-    . f e 2 2 2 2 e e f f f f . . . 
-    . f 2 e f f f f 2 2 2 e f . . . 
-    . f f f e e e f f f f f f f . . 
-    . f e e 4 4 f b e 4 4 e f f . . 
-    . . f e d d f 1 4 d 4 e e f . . 
-    . . . f d d d d 4 e e e f . . . 
-    . . . f e 4 4 4 e e f f . . . . 
-    . . . f 2 2 2 e d d 4 . . . . . 
-    . . . f 2 2 2 e d d e . . . . . 
-    . . . f 5 5 4 f e e f . . . . . 
-    . . . . f f f f f f . . . . . . 
-    . . . . . . f f f . . . . . . . 
-    `, img`
-    . . . . . . . . . . . . . . . . 
-    . . . . f f f f f f . . . . . . 
-    . . . f 2 f e e e e f f . . . . 
-    . . f 2 2 2 f e e e e f f . . . 
-    . . f e e e e f f e e e f . . . 
-    . f e 2 2 2 2 e e f f f f . . . 
-    . f 2 e f f f f 2 2 2 e f . . . 
-    . f f f e e e f f f f f f f . . 
-    . f e e 4 4 f b e 4 4 e f f . . 
-    . . f e d d f 1 4 d 4 e e f . . 
-    . . . f d d d e e e e e f . . . 
-    . . . f e 4 e d d 4 f . . . . . 
-    . . . f 2 2 e d d e f . . . . . 
-    . . f f 5 5 f e e f f f . . . . 
-    . . f f f f f f f f f f . . . . 
-    . . . f f f . . . f f . . . . . 
-    `, img`
-    . . . . f f f f f f . . . . . . 
-    . . . f 2 f e e e e f f . . . . 
-    . . f 2 2 2 f e e e e f f . . . 
-    . . f e e e e f f e e e f . . . 
-    . f e 2 2 2 2 e e f f f f . . . 
-    . f 2 e f f f f 2 2 2 e f . . . 
-    . f f f e e e f f f f f f f . . 
-    . f e e 4 4 f b e 4 4 e f f . . 
-    . . f e d d f 1 4 d 4 e e f . . 
-    . . . f d d d d 4 e e e f . . . 
-    . . . f e 4 4 4 e e f f . . . . 
-    . . . f 2 2 2 e d d 4 . . . . . 
-    . . . f 2 2 2 e d d e . . . . . 
-    . . . f 5 5 4 f e e f . . . . . 
-    . . . . f f f f f f . . . . . . 
-    . . . . . . f f f . . . . . . . 
-    `, img`
-    . . . . . . . . . . . . . . . . 
-    . . . . f f f f f f . . . . . . 
-    . . . f 2 f e e e e f f . . . . 
-    . . f 2 2 2 f e e e e f f . . . 
-    . . f e e e e f f e e e f . . . 
-    . f e 2 2 2 2 e e f f f f . . . 
-    . f 2 e f f f f 2 2 2 e f . . . 
-    . f f f e e e f f f f f f f . . 
-    . f e e 4 4 f b e 4 4 e f f . . 
-    . . f e d d f 1 4 d 4 e e f . . 
-    . . . f d d d d 4 e e e f . . . 
-    . . . f e 4 4 4 e d d 4 . . . . 
-    . . . f 2 2 2 2 e d d e . . . . 
-    . . f f 5 5 4 4 f e e f . . . . 
-    . . f f f f f f f f f f . . . . 
-    . . . f f f . . . f f . . . . . 
-    `],
-    150,
-    character.rule(Predicate.MovingLeft)
-)
-character.loopFrames(
-    sprHero,
-    [img`
-    . . . . . f f 4 4 f f . . . . . 
-    . . . . f 5 4 5 5 4 5 f . . . . 
-    . . . f e 4 5 5 5 5 4 e f . . . 
-    . . f b 3 e 4 4 4 4 e 3 b f . . 
-    . . f 3 3 3 3 3 3 3 3 3 3 f . . 
-    . f 3 3 e b 3 e e 3 b e 3 3 f . 
-    . f 3 3 f f e e e e f f 3 3 f . 
-    . f b b f b f e e f b f b b f . 
-    . f b b e 1 f 4 4 f 1 e b b f . 
-    f f b b f 4 4 4 4 4 4 f b b f f 
-    f b b f f f e e e e f f f b b f 
-    . f e e f b d d d d b f e e f . 
-    . . e 4 c d d d d d d c 4 e . . 
-    . . e f b d b d b d b b f e . . 
-    . . . f f 1 d 1 d 1 d f f . . . 
-    . . . . . f f b b f f . . . . . 
-    `, img`
-    . . . . . . . f f . . . . . . . 
-    . . . . . f f 4 4 f f . . . . . 
-    . . . . f 5 4 5 5 4 5 f . . . . 
-    . . . f e 4 5 5 5 5 4 e f . . . 
-    . . f b 3 e 4 4 4 4 e 3 b f . . 
-    . f e 3 3 3 3 3 3 3 3 3 3 e f . 
-    . f 3 3 e b 3 e e 3 b e 3 3 f . 
-    . f b 3 f f e e e e f f 3 b f . 
-    f f b b f b f e e f b f b b f f 
-    f b b b e 1 f 4 4 f 1 e b b b f 
-    . f b b e e 4 4 4 4 4 f b b f . 
-    . . f 4 4 4 e d d d b f e f . . 
-    . . f e 4 4 e d d d d c 4 e . . 
-    . . . f e e d d b d b b f e . . 
-    . . . f f 1 d 1 d 1 1 f f . . . 
-    . . . . . f f f b b f . . . . . 
-    `, img`
-    . . . . . f f 4 4 f f . . . . . 
-    . . . . f 5 4 5 5 4 5 f . . . . 
-    . . . f e 4 5 5 5 5 4 e f . . . 
-    . . f b 3 e 4 4 4 4 e 3 b f . . 
-    . . f 3 3 3 3 3 3 3 3 3 3 f . . 
-    . f 3 3 e b 3 e e 3 b e 3 3 f . 
-    . f 3 3 f f e e e e f f 3 3 f . 
-    . f b b f b f e e f b f b b f . 
-    . f b b e 1 f 4 4 f 1 e b b f . 
-    f f b b f 4 4 4 4 4 4 f b b f f 
-    f b b f f f e e e e f f f b b f 
-    . f e e f b d d d d b f e e f . 
-    . . e 4 c d d d d d d c 4 e . . 
-    . . e f b d b d b d b b f e . . 
-    . . . f f 1 d 1 d 1 d f f . . . 
-    . . . . . f f b b f f . . . . . 
-    `, img`
-    . . . . . . . f f . . . . . . . 
-    . . . . . f f 4 4 f f . . . . . 
-    . . . . f 5 4 5 5 4 5 f . . . . 
-    . . . f e 4 5 5 5 5 4 e f . . . 
-    . . f b 3 e 4 4 4 4 e 3 b f . . 
-    . f e 3 3 3 3 3 3 3 3 3 3 e f . 
-    . f 3 3 e b 3 e e 3 b e 3 3 f . 
-    . f b 3 f f e e e e f f 3 b f . 
-    f f b b f b f e e f b f b b f f 
-    f b b b e 1 f 4 4 f 1 e b b b f 
-    . f b b f 4 4 4 4 4 e e b b f . 
-    . . f e f b d d d e 4 4 4 f . . 
-    . . e 4 c d d d d e 4 4 e f . . 
-    . . e f b b d b d d e e f . . . 
-    . . . f f 1 1 d 1 d 1 f f . . . 
-    . . . . . f b b f f f . . . . . 
-    `],
-    150,
-    character.rule(Predicate.NotMoving)
-)
-character.loopFrames(
-    sprPriness,
-    [img`
-    . . . . . . f f f f 4 4 f . . . 
-    . . . . f f b f 5 4 5 5 4 f . . 
-    . . . f b 3 3 e 4 5 5 5 5 f . . 
-    . . f b 3 3 3 3 e 4 4 4 e f . . 
-    . . f 3 3 3 3 3 3 3 3 3 3 f . . 
-    . . f 3 3 3 3 e b 3 e e 3 3 f . 
-    . . f 3 3 3 3 f f e e e 3 3 f . 
-    . . f b b b b f b f e e e 3 f . 
-    . . f b b b b e 1 f 4 4 e f . . 
-    . f f b b b b f 4 4 4 4 f . . . 
-    . f b b b b f f f e e e f . . . 
-    . . f b b f 4 4 e d d d f . . . 
-    . . . f f e 4 4 e d d d f . . . 
-    . . . . f b e e b d b d b f . . 
-    . . . . f f d 1 d 1 d 1 f f . . 
-    . . . . . . f f b b f f . . . . 
-    `, img`
-    . . . . . . . . . . . . . . . . 
-    . . . . . . f f f f 4 4 f . . . 
-    . . . . f f b f 5 4 5 5 4 f . . 
-    . . . f b 3 3 e 4 5 5 5 5 f . . 
-    . . f b 3 3 3 3 e 4 4 4 e f . . 
-    . . f 3 3 3 3 3 3 3 3 3 3 3 f . 
-    . . f 3 3 3 3 e b 3 e e 3 3 f . 
-    . . f 3 3 3 3 f f e e e 3 3 f . 
-    . f f b b b b f b f e e e f . . 
-    . f b b b b b e 1 f 4 4 e . . . 
-    . f b b b b b f 4 4 4 4 f . . . 
-    . . f b b b 4 4 e d d d f . . . 
-    . . . f f f 4 4 e d d d f . . . 
-    . . . f b b e e b b d d d f . . 
-    . . . . f b d d 1 d 1 d b f . . 
-    . . . . . f f f b b f f f . . . 
-    `],
-    150,
-    character.rule(Predicate.MovingRight)
-)
-character.loopFrames(
-    sprPriness,
-    [img`
-    . . . f 4 4 f f f f . . . . . . 
-    . . f 4 5 5 4 5 f b f f . . . . 
-    . . f 5 5 5 5 4 e 3 3 b f . . . 
-    . . f e 4 4 4 e 3 3 3 3 b f . . 
-    . . f 3 3 3 3 3 3 3 3 3 3 f . . 
-    . f 3 3 e e 3 b e 3 3 3 3 f . . 
-    . f 3 3 e e e f f 3 3 3 3 f . . 
-    . f 3 e e e f b f b b b b f . . 
-    . . f e 4 4 f 1 e b b b b f . . 
-    . . . f 4 4 4 4 f b b b b f f . 
-    . . . f e e e f f f b b b b f . 
-    . . . f d d d e 4 4 f b b f . . 
-    . . . f d d d e 4 4 e f f . . . 
-    . . f b d b d b e e b f . . . . 
-    . . f f 1 d 1 d 1 d f f . . . . 
-    . . . . f f b b f f . . . . . . 
-    `, img`
-    . . . . . . . . . . . . . . . . 
-    . . . f 4 4 f f f f . . . . . . 
-    . . f 4 5 5 4 5 f b f f . . . . 
-    . . f 5 5 5 5 4 e 3 3 b f . . . 
-    . . f e 4 4 4 e 3 3 3 3 b f . . 
-    . f 3 3 3 3 3 3 3 3 3 3 3 f . . 
-    . f 3 3 e e 3 b e 3 3 3 3 f . . 
-    . f 3 3 e e e f f 3 3 3 3 f . . 
-    . . f e e e f b f b b b b f f . 
-    . . . e 4 4 f 1 e b b b b b f . 
-    . . . f 4 4 4 4 f b b b b b f . 
-    . . . f d d d e 4 4 b b b f . . 
-    . . . f d d d e 4 4 f f f . . . 
-    . . f d d d b b e e b b f . . . 
-    . . f b d 1 d 1 d d b f . . . . 
-    . . . f f f b b f f f . . . . . 
-    `],
-    150,
-    character.rule(Predicate.MovingLeft)
-)
-character.loopFrames(
-    sprPriness,
-    [img`
-    . . . . . f f 4 4 f f . . . . . 
-    . . . . f 5 4 5 5 4 5 f . . . . 
-    . . . f e 4 5 5 5 5 4 e f . . . 
-    . . f b 3 e 4 4 4 4 e 3 b f . . 
-    . . f 3 3 3 3 3 3 3 3 3 3 f . . 
-    . f 3 3 e b 3 e e 3 b e 3 3 f . 
-    . f 3 3 f f e e e e f f 3 3 f . 
-    . f b b f b f e e f b f b b f . 
-    . f b b e 1 f 4 4 f 1 e b b f . 
-    f f b b f 4 4 4 4 4 4 f b b f f 
-    f b b f f f e e e e f f f b b f 
-    . f e e f b d d d d b f e e f . 
-    . . e 4 c d d d d d d c 4 e . . 
-    . . e f b d b d b d b b f e . . 
-    . . . f f 1 d 1 d 1 d f f . . . 
-    . . . . . f f b b f f . . . . . 
-    `, img`
-    . . . . . . . f f . . . . . . . 
-    . . . . . f f 4 4 f f . . . . . 
-    . . . . f 5 4 5 5 4 5 f . . . . 
-    . . . f e 4 5 5 5 5 4 e f . . . 
-    . . f b 3 e 4 4 4 4 e 3 b f . . 
-    . f e 3 3 3 3 3 3 3 3 3 3 e f . 
-    . f 3 3 e b 3 e e 3 b e 3 3 f . 
-    . f b 3 f f e e e e f f 3 b f . 
-    f f b b f b f e e f b f b b f f 
-    f b b b e 1 f 4 4 f 1 e b b b f 
-    . f b b e e 4 4 4 4 4 f b b f . 
-    . . f 4 4 4 e d d d b f e f . . 
-    . . f e 4 4 e d d d d c 4 e . . 
-    . . . f e e d d b d b b f e . . 
-    . . . f f 1 d 1 d 1 1 f f . . . 
-    . . . . . f f f b b f . . . . . 
-    `, img`
-    . . . . . f f 4 4 f f . . . . . 
-    . . . . f 5 4 5 5 4 5 f . . . . 
-    . . . f e 4 5 5 5 5 4 e f . . . 
-    . . f b 3 e 4 4 4 4 e 3 b f . . 
-    . . f 3 3 3 3 3 3 3 3 3 3 f . . 
-    . f 3 3 e b 3 e e 3 b e 3 3 f . 
-    . f 3 3 f f e e e e f f 3 3 f . 
-    . f b b f b f e e f b f b b f . 
-    . f b b e 1 f 4 4 f 1 e b b f . 
-    f f b b f 4 4 4 4 4 4 f b b f f 
-    f b b f f f e e e e f f f b b f 
-    . f e e f b d d d d b f e e f . 
-    . . e 4 c d d d d d d c 4 e . . 
-    . . e f b d b d b d b b f e . . 
-    . . . f f 1 d 1 d 1 d f f . . . 
-    . . . . . f f b b f f . . . . . 
-    `, img`
-    . . . . . . . f f . . . . . . . 
-    . . . . . f f 4 4 f f . . . . . 
-    . . . . f 5 4 5 5 4 5 f . . . . 
-    . . . f e 4 5 5 5 5 4 e f . . . 
-    . . f b 3 e 4 4 4 4 e 3 b f . . 
-    . f e 3 3 3 3 3 3 3 3 3 3 e f . 
-    . f 3 3 e b 3 e e 3 b e 3 3 f . 
-    . f b 3 f f e e e e f f 3 b f . 
-    f f b b f b f e e f b f b b f f 
-    f b b b e 1 f 4 4 f 1 e b b b f 
-    . f b b f 4 4 4 4 4 e e b b f . 
-    . . f e f b d d d e 4 4 4 f . . 
-    . . e 4 c d d d d e 4 4 e f . . 
-    . . e f b b d b d d e e f . . . 
-    . . . f f 1 1 d 1 d 1 f f . . . 
-    . . . . . f b b f f f . . . . . 
-    `],
-    150,
-    character.rule(Predicate.MovingDown)
-)
-character.loopFrames(
-    sprPriness,
-    [img`
-    . . . . . f f 4 4 f f . . . . . 
-    . . . . f 5 4 5 5 4 5 f . . . . 
-    . . . f e 3 3 3 3 3 3 e f . . . 
-    . . f b 3 3 3 3 3 3 3 3 b f . . 
-    . . f 3 3 3 3 3 3 3 3 3 3 f . . 
-    . f 3 3 3 3 3 3 3 3 3 3 3 3 f . 
-    . f b 3 3 3 3 3 3 3 3 3 3 b f . 
-    . f b b 3 3 3 3 3 3 3 3 b b f . 
-    . f b b b b b b b b b b b b f . 
-    f c b b b b b b b b b b b b c f 
-    f b b b b b b b b b b b b b b f 
-    . f c c b b b b b b b b c c f . 
-    . . e 4 c f f f f f f c 4 e . . 
-    . . e f b d b d b d b b f e . . 
-    . . . f f 1 d 1 d 1 d f f . . . 
-    . . . . . f f b b f f . . . . . 
-    `, img`
-    . . . . . . . . . . . . . . . . 
-    . . . . . f f 4 4 f f . . . . . 
-    . . . . f 5 4 5 5 4 5 f . . . . 
-    . . . f e 3 3 3 3 3 3 e f . . . 
-    . . f b 3 3 3 3 3 3 3 3 b f . . 
-    . . f 3 3 3 3 3 3 3 3 3 3 f . . 
-    . f b 3 3 3 3 3 3 3 3 3 3 b f . 
-    . f b b 3 3 3 3 3 3 3 3 b b f . 
-    . f b b b b b b b b b b b b f . 
-    f c b b b b b b b b b b b b f . 
-    f b b b b b b b b b b b b c f . 
-    f f b b b b b b b b b b c f . . 
-    . f c c c f f f f f f f e c . . 
-    . . . f b b d b d d e 4 4 e . . 
-    . . . f f 1 1 d 1 d e e f . . . 
-    . . . . . f b b f f f . . . . . 
-    `, img`
-    . . . . . f f 4 4 f f . . . . . 
-    . . . . f 5 4 5 5 4 5 f . . . . 
-    . . . f e 3 3 3 3 3 3 e f . . . 
-    . . f b 3 3 3 3 3 3 3 3 b f . . 
-    . . f 3 3 3 3 3 3 3 3 3 3 f . . 
-    . f 3 3 3 3 3 3 3 3 3 3 3 3 f . 
-    . f b 3 3 3 3 3 3 3 3 3 3 b f . 
-    . f b b 3 3 3 3 3 3 3 3 b b f . 
-    . f b b b b b b b b b b b b f . 
-    f c b b b b b b b b b b b b c f 
-    f b b b b b b b b b b b b b b f 
-    . f c c b b b b b b b b c c f . 
-    . . e 4 c f f f f f f c 4 e . . 
-    . . e f b d b d b d b b f e . . 
-    . . . f f 1 d 1 d 1 d f f . . . 
-    . . . . . f f b b f f . . . . . 
-    `, img`
-    . . . . . . . . . . . . . . . . 
-    . . . . . f f 4 4 f f . . . . . 
-    . . . . f 5 4 5 5 4 5 f . . . . 
-    . . . f e 3 3 3 3 3 3 e f . . . 
-    . . f b 3 3 3 3 3 3 3 3 b f . . 
-    . . f 3 3 3 3 3 3 3 3 3 3 f . . 
-    . f b 3 3 3 3 3 3 3 3 3 3 b f . 
-    . f b b 3 3 3 3 3 3 3 3 b b f . 
-    . f b b b b b b b b b b b b f . 
-    . f b b b b b b b b b b b b c f 
-    . f c b b b b b b b b b b b b f 
-    . . f c b b b b b b b b b b f f 
-    . . c e f f f f f f f c c c f . 
-    . . e 4 4 e d d b d b b f . . . 
-    . . . f e e d 1 d 1 1 f f . . . 
-    . . . . . f f f b b f . . . . . 
-    `],
-    150,
-    character.rule(Predicate.MovingUp)
-)
-// forever(function () {
-//     if (controller.A.isPressed()) {
-//         if (controller.left.isPressed()) {
-//             character.setCharacterState(sprHero, character.rule(Predicate.MovingLeft))
-//         } else if (controller.up.isPressed()) {
-//             character.setCharacterState(sprHero, character.rule(Predicate.MovingUp))
-//         } else if (controller.right.isPressed()) {
-//             character.setCharacterState(sprHero, character.rule(Predicate.MovingRight))
-//         } else if (controller.down.isPressed()) {
-//             character.setCharacterState(sprHero, character.rule(Predicate.MovingDown))
-//         }
-//     } else {
-//         if (controller.left.isPressed()) {
-//             character.setCharacterState(sprHero, character.rule(Predicate.MovingLeft))
-//         } else if (controller.up.isPressed()) {
-//             character.setCharacterState(sprHero, character.rule(Predicate.MovingUp))
-//         } else if (controller.right.isPressed()) {
-//             character.setCharacterState(sprHero, character.rule(Predicate.MovingRight))
-//         } else if (controller.down.isPressed()) {
-//             character.setCharacterState(sprHero, character.rule(Predicate.MovingDown))
-//         }
-//     }
-// })
-
